不保留小数

See merge request industrial/smartindustry!325

+++ inventorymanage/src/main/java/com/smartindustry/inventory/vo/MaterialInventoryVO.java
@@ -78,7 +78,7 @@ public class MaterialInventoryVO implements Serializable {
         }
         vo.setLlimit(bo.getLowerLimit());
         if (bo.getLowerLimit() != null) {
-            vo.setLlimitmunit(bo.getLowerLimit() + munit);
+            vo.setLlimitmunit(bo.getLowerLimit().longValue() + munit);
         }
         vo.setStatus(bo.getStatus());
         vo.setLnum(bo.getLockNum());
